Refactor the client (browser) javascript
Instead of sending the login credentials right away after SockJS connects, now
first wait for a connected=true message from the funnel server. That way no
messages are lost.

Also, improve on the ``send`` function, introduce a ``status`` function,
and a ``process`` function - the former two will need a lot of improvement
in the future.