the client now tells which host, port the funnel to connect
With this change, the host and port of the irc server (or znc bouncer)
are configured in the browser client which instructs the funnel what to
connect to.

While this makes the server code more flexible and generic, it also makes
it less secure, or at least exploitable - it's  basically an open proxy.
I hope noone finds where I run my own instance :)

BTW, since nodejs uses callbacks, a 4 line change in server.js,
makes it look like a full rewrite of it. It's a lie!
(maybe I should stop indenting callbacks???)

Also, I might revert this sometime